Fixed database connection leak
The database was leaking connections when constructing a RequestState
and the database was get a connection and then the graph database would
fail on getting a connection.  At that point the database connection was
never cleaned up.  The solution implemented was to wrap getting the
connections in a try/catch and closing the connections if an exception
is thrown.